fix: stuck agent creation with local backend
The issue occurs when creating agent and set the backend to local
which cause the terminal hang and not responding. This is because
there is a process to call Letta Cloud to sync the memFS which
suppose to be happen if the backend is set to API. This commit
fixes it by adding a flag check if backend has remoteMemfs
capabilities set to true. If false, it will bypass the sync.
